Updating a note on using AWVERIFY records
I tested this with a customer, but the note on using AWVERIFY records is confusing. Since Traffic Manager does not allow for the root domain to be mapped to it, having an example saying "create an awverify record from 'awverify.contoso.com' to 'contoso.trafficmanager.net' is confusing. I updated it to show an example of preemptively binding the www subdomain to a web app.